public/kcl-samples/requirements.txt
pillow==11.1.0
requests==2.32.3
zoo-kcl==0.3.38